requirements.txt
google-generativeai
pdfplumber
python-docx
